Sign In view debug
