+++ MessengerApp.xcodeproj/xcuserdata/gentjanmanuka.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "A516CCB7-53F2-476C-BA9F-D60E4DABF357"
+   type = "1"
+   version = "2.0">
+</Bucket>
